+++ src/breakpoints.h
@@ -419,7 +419,7 @@ public:
   void set_message(unsigned int b,string &);
   void clear_all(Processor *c);
   void clear_all_set_by_user(Processor *c);
-  void clear_all_register(Processor *c,unsigned int address=-1);
+  void clear_all_register(Processor *c,gint64  address=-1);
   void initialize_breakpoints(unsigned int memory_size);
   instruction *find_previous(Processor *cpu, 
 			     unsigned int address,
